Moved application code to single file.
